Add new items; couple bugs
Adding new items works for the database structure. I need to work on editing the fragment stack because the barcode gets stuck on the screen when the user goes back in some cases. Attempted to use actual phone and there is an issue connecting to the server with the device.