Home and remove fragment
The home fragment now displays the 10 most recently edited items. The remove fragment redirects to add new item or remove item quantity depending if the item is in the database. The locations on the remove screen only list the valid locations of that item. Might need to make a change object for a better way to pass data. Working on only making one post request to reduce volley string requests. If this is successful, I will need to change the add volley and PHP files.